Run a version that is up-to-date with the README
By default generate a job configuration that executes the up-to-date version of the code.

Or alternatively just release a new version. The current release is from February and predates the `minimumChangeThreshold` setting documented in the README, so [the current situation wastes people's time troubleshooting why that setting doesn't work](https://github.com/hashicorp/nextjs-bundle-analysis/issues/25). :-)